fix(Excel): 修正Top属性的数据类型转换
将ExcelApplication中Top属性的getter从Convert.ToSingle改为Convert.ToInt32，
确保返回整数类型值而不是浮点数类型值。

feat(Word): 添加WdMailMergeDestination枚举的中文文档注释

为WdMailMergeDestination枚举及其所有成员添加详细的中文文档注释，
包括邮件合并目的地的各种选项说明。

+++ MudTools.OfficeInterop.Excel/Core/Core/Imps/ExcelApplication.cs
@@ -1413,7 +1413,7 @@ internal partial class ExcelApplication : IExcelApplication
     /// </summary>
     public int Top
     {
-        get => Convert.ToSingle(_application?.Top);
+        get => Convert.ToInt32(_application?.Top);
         set
         {
             if (_application != null)
